Move tests to lowercase, add Scrutinizer support, break UriHelper test cases, add test for fail condition
