+++ .scrutinizer.yml
@@ -1,0 +1,10 @@
+imports:
+    - php
+
+tools:
+    external_code_coverage: true
+
+filter:
+    excluded_paths:
+        - tests/*
+        - vendor/*
