+++ .travis.yml
@@ -15,5 +15,8 @@ before_script:
   - composer update
 
 script:
-  - ./vendor/bin/phpunit
-  - ./vendor/bin/phpcs -p --report=full --extensions=php --standard=.travis/phpcs/Joomla/ruleset.xml src/
+  - mkdir -p .travis/logs
+  - vendor/bin/phpunit --coverage-clover .travis/logs/clover.xml
+  - wget https://scrutinizer-ci.com/ocular.phar
+  - php ocular.phar code-coverage:upload --format=php-clover .travis/logs/clover.xml
+  - vendor/bin/phpcs -p --report=full --extensions=php --standard=.travis/phpcs/Joomla/ruleset.xml src/
