+++ phpunit.xml.dist
@@ -2,7 +2,7 @@
 <phpunit bootstrap="vendor/autoload.php" colors="false">
 	<testsuites>
 		<testsuite name="Unit">
-			<directory>Tests</directory>
+			<directory>tests</directory>
 		</testsuite>
 	</testsuites>
 </phpunit>
