+++ tests/UriImmutableTest.php
